+++ tests/testthat/test-5-methods.R
@@ -676,7 +676,7 @@ test_that("Effect size summary functions work", {
 #### creating / updating the test settings ####
 if(FALSE){
 
-  saved_files <- paste0("fit_", 1:16, ".RDS")
+  saved_files <- paste0("fit_", 1:17, ".RDS")
   saved_fits  <- list()
   for(i in seq_along(saved_files)){
     saved_fits[[i]] <- readRDS(file = file.path("tests/results/fits", saved_files[i]))